Ignore one HashMap test for foldBack
Ignore one of the HashMap tests for the foldBack function which'd been ported over
from the FSharp.Core tests for Map. The test relies on the order in which the elements
are traversed, which isn't really sensible for a hash-based (unordered) map. This test
could be re-enabled in the future if we define a custom key type to use for the test
whose hash function won't change across systems / process runs, rather than using
strings as it does now.